+++ src/Patches/SoulDetachmentTurnControlPatch.cs
@@ -20,7 +20,7 @@ internal static class SoulDetachmentTurnControlPatch
 	private static void Postfix(
 		ICombatState combatState,
 		CombatSide side,
-		IEnumerable<Creature> creatures,
+		IEnumerable<Creature> participants,
 		ref Task __result)
 	{
 		__result = RunAfterHook(__result, combatState, side);